+++ zenoh/src/api/session.rs
@@ -17,7 +17,6 @@ use std::{
     collections::HashMap,
     convert::TryInto,
     fmt,
-    future::{IntoFuture, Ready},
     ops::Deref,
     sync::{
         atomic::{AtomicU16, Ordering},
@@ -33,7 +32,7 @@ use uhlc::HLC;
 use zenoh_buffers::ZBuf;
 use zenoh_collections::SingleOrVec;
 use zenoh_config::{unwrap_or_default, wrappers::ZenohId};
-use zenoh_core::{zconfigurable, zread, Resolvable, Resolve, ResolveClosure, ResolveFuture, Wait};
+use zenoh_core::{zconfigurable, zread, Resolve, ResolveClosure, ResolveFuture, Wait};
 #[cfg(feature = "unstable")]
 use zenoh_protocol::network::{
     declare::{DeclareToken, SubscriberId, TokenId, UndeclareToken},
@@ -68,43 +67,45 @@ use zenoh_result::ZResult;
 use zenoh_shm::api::client_storage::ShmClientStorage;
 use zenoh_task::TaskController;
 
-use super::{
-    admin,
-    builders::{
-        publisher::{
-            PublicationBuilderDelete, PublicationBuilderPut, PublisherBuilder,
-            SessionDeleteBuilder, SessionPutBuilder,
-        },
-        queryable::QueryableBuilder,
-        subscriber::SubscriberBuilder,
-    },
-    bytes::ZBytes,
-    encoding::Encoding,
-    handlers::{Callback, DefaultHandler},
-    info::SessionInfo,
-    key_expr::{KeyExpr, KeyExprInner},
-    publisher::{Priority, PublisherState},
-    query::{
-        ConsolidationMode, QueryConsolidation, QueryState, QueryTarget, Reply, SessionGetBuilder,
-    },
-    queryable::{Query, QueryInner, QueryableState},
-    sample::{DataInfo, DataInfoIntoSample, Locality, QoS, Sample, SampleKind},
-    selector::Selector,
-    subscriber::{SubscriberKind, SubscriberState},
-    value::Value,
-    Id,
-};
 #[cfg(feature = "unstable")]
-use super::{
+use crate::api::selector::ZenohParameters;
+#[cfg(feature = "unstable")]
+use crate::api::{
     liveliness::{Liveliness, LivelinessTokenState},
     publisher::Publisher,
     publisher::{MatchingListenerState, MatchingStatus},
     query::LivelinessQueryState,
     sample::SourceInfo,
 };
-#[cfg(feature = "unstable")]
-use crate::api::selector::ZenohParameters;
 use crate::{
+    api::{
+        admin,
+        builders::{
+            publisher::{
+                PublicationBuilderDelete, PublicationBuilderPut, PublisherBuilder,
+                SessionDeleteBuilder, SessionPutBuilder,
+            },
+            queryable::QueryableBuilder,
+            session::OpenBuilder,
+            subscriber::SubscriberBuilder,
+        },
+        bytes::ZBytes,
+        encoding::Encoding,
+        handlers::{Callback, DefaultHandler},
+        info::SessionInfo,
+        key_expr::{KeyExpr, KeyExprInner},
+        publisher::{Priority, PublisherState},
+        query::{
+            ConsolidationMode, QueryConsolidation, QueryState, QueryTarget, Reply,
+            SessionGetBuilder,
+        },
+        queryable::{Query, QueryInner, QueryableState},
+        sample::{DataInfo, DataInfoIntoSample, Locality, QoS, Sample, SampleKind},
+        selector::Selector,
+        subscriber::{SubscriberKind, SubscriberState},
+        value::Value,
+        Id,
+    },
     net::{
         primitives::Primitives,
         routing::dispatcher::face::Face,
@@ -2879,146 +2880,5 @@ where
     TryIntoConfig: std::convert::TryInto<crate::config::Config> + Send + 'static,
     <TryIntoConfig as std::convert::TryInto<crate::config::Config>>::Error: std::fmt::Debug,
 {
-    OpenBuilder {
-        config,
-        #[cfg(feature = "shared-memory")]
-        shm_clients: None,
-    }
-}
-
-/// A builder returned by [`open`] used to open a zenoh [`Session`].
-///
-/// # Examples
-/// ```
-/// # #[tokio::main]
-/// # async fn main() {
-///
-/// let session = zenoh::open(zenoh::Config::default()).await.unwrap();
-/// # }
-/// ```
-#[must_use = "Resolvables do nothing unless you resolve them using `.await` or `zenoh::Wait::wait`"]
-pub struct OpenBuilder<TryIntoConfig>
-where
-    TryIntoConfig: std::convert::TryInto<crate::config::Config> + Send + 'static,
-    <TryIntoConfig as std::convert::TryInto<crate::config::Config>>::Error: std::fmt::Debug,
-{
-    config: TryIntoConfig,
-    #[cfg(feature = "shared-memory")]
-    shm_clients: Option<Arc<ShmClientStorage>>,
-}
-
-#[cfg(feature = "shared-memory")]
-impl<TryIntoConfig> OpenBuilder<TryIntoConfig>
-where
-    TryIntoConfig: std::convert::TryInto<crate::config::Config> + Send + 'static,
-    <TryIntoConfig as std::convert::TryInto<crate::config::Config>>::Error: std::fmt::Debug,
-{
-    pub fn with_shm_clients(mut self, shm_clients: Arc<ShmClientStorage>) -> Self {
-        self.shm_clients = Some(shm_clients);
-        self
-    }
-}
-
-impl<TryIntoConfig> Resolvable for OpenBuilder<TryIntoConfig>
-where
-    TryIntoConfig: std::convert::TryInto<crate::config::Config> + Send + 'static,
-    <TryIntoConfig as std::convert::TryInto<crate::config::Config>>::Error: std::fmt::Debug,
-{
-    type To = ZResult<Session>;
-}
-
-impl<TryIntoConfig> Wait for OpenBuilder<TryIntoConfig>
-where
-    TryIntoConfig: std::convert::TryInto<crate::config::Config> + Send + 'static,
-    <TryIntoConfig as std::convert::TryInto<crate::config::Config>>::Error: std::fmt::Debug,
-{
-    fn wait(self) -> <Self as Resolvable>::To {
-        let config: crate::config::Config = self
-            .config
-            .try_into()
-            .map_err(|e| zerror!("Invalid Zenoh configuration {:?}", &e))?;
-        Session::new(
-            config,
-            #[cfg(feature = "shared-memory")]
-            self.shm_clients,
-        )
-        .wait()
-    }
-}
-
-impl<TryIntoConfig> IntoFuture for OpenBuilder<TryIntoConfig>
-where
-    TryIntoConfig: std::convert::TryInto<crate::config::Config> + Send + 'static,
-    <TryIntoConfig as std::convert::TryInto<crate::config::Config>>::Error: std::fmt::Debug,
-{
-    type Output = <Self as Resolvable>::To;
-    type IntoFuture = Ready<<Self as Resolvable>::To>;
-
-    fn into_future(self) -> Self::IntoFuture {
-        std::future::ready(self.wait())
-    }
-}
-
-/// Initialize a Session with an existing Runtime.
-/// This operation is used by the plugins to share the same Runtime as the router.
-#[zenoh_macros::internal]
-pub fn init(runtime: Runtime) -> InitBuilder {
-    InitBuilder {
-        runtime,
-        aggregated_subscribers: vec![],
-        aggregated_publishers: vec![],
-    }
-}
-
-/// A builder returned by [`init`] and used to initialize a Session with an existing Runtime.
-#[must_use = "Resolvables do nothing unless you resolve them using `.await` or `zenoh::Wait::wait`"]
-#[doc(hidden)]
-#[zenoh_macros::internal]
-pub struct InitBuilder {
-    runtime: Runtime,
-    aggregated_subscribers: Vec<OwnedKeyExpr>,
-    aggregated_publishers: Vec<OwnedKeyExpr>,
-}
-
-#[zenoh_macros::internal]
-impl InitBuilder {
-    #[inline]
-    pub fn aggregated_subscribers(mut self, exprs: Vec<OwnedKeyExpr>) -> Self {
-        self.aggregated_subscribers = exprs;
-        self
-    }
-
-    #[inline]
-    pub fn aggregated_publishers(mut self, exprs: Vec<OwnedKeyExpr>) -> Self {
-        self.aggregated_publishers = exprs;
-        self
-    }
-}
-
-#[zenoh_macros::internal]
-impl Resolvable for InitBuilder {
-    type To = ZResult<Session>;
-}
-
-#[zenoh_macros::internal]
-impl Wait for InitBuilder {
-    fn wait(self) -> <Self as Resolvable>::To {
-        Ok(Session::init(
-            self.runtime,
-            self.aggregated_subscribers,
-            self.aggregated_publishers,
-            false,
-        )
-        .wait())
-    }
-}
-
-#[zenoh_macros::internal]
-impl IntoFuture for InitBuilder {
-    type Output = <Self as Resolvable>::To;
-    type IntoFuture = Ready<<Self as Resolvable>::To>;
-
-    fn into_future(self) -> Self::IntoFuture {
-        std::future::ready(self.wait())
-    }
+    OpenBuilder::new(config)
 }
